feat(desktopCharting): SAV-773: Add complex chart graph importer validation (#8145)
* Prevent complex chart surveys to have visualisation config

* Create unit test coverage

* Create xlsx file for unit case

* Add visualisation config to simple chart survey for coverage

* Reverse check order and improve unit test check

* refactor(tests): SAV-773: Split programs importer test file (#8151)

* Refactor into smaller files

* Remove unneded table truncates per file and declare them inline on hook, also ditch doing that on afterAll hook
