+++ packages/central-server/app/admin/programImporter/programDataElementValidation.js
@@ -95,6 +95,12 @@ export function validateProgramDataElementRecords(
     const { validationCriteria = '' } = surveyScreenComponentRecord.values;
 
     try {
+      if (
+        visualisationConfig &&
+        [SURVEY_TYPES.COMPLEX_CHART, SURVEY_TYPES.COMPLEX_CHART_CORE].includes(surveyType)
+      ) {
+        throw new Error('Visualisation config is not allowed for complex charts');
+      }
       if (config.validateQuestionConfigs.enabled) {
         validateVisualisationConfig(visualisationConfig, validationCriteria);
       }
